[sync] Migrate Olmo and Olmo2 to the Transformers modeling backend (#48100)
Ported the Olmo/Olmo2 Transformers backend migration: removed native Olmo/Olmo2 model files, added Olmo3-native module, and updated the model registry so Olmo and Olmo2 use the Transformers backend while Olmo3 uses the new native module. Syntax check passed.

Upstream-vLLM: b83be00cddf03387bf740acfe6b9ca07ec1a3c08
